+++ src/app/routing/ui/Routing.tsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from 'react-router-dom';
-import { ROUTES } from '@shared/router';
-import { useRedirect } from '@shared/redirect';
+import { ROUTES } from '@shared/consts';
+import { useRedirect } from '@shared/lib';
 import { AppLayout } from '@widgets/layout';
 import NotFoundPage from '@pages/not-found';
 import LoginPage from '@pages/login-page';
