+++ src/entities/session/lib/tokenLocalStorageHandlers.ts
@@ -2,8 +2,8 @@ import {
   getFromLocalStorage,
   removeFromLocalStorage,
   setToLocalStorage,
-} from '@shared/webStorage';
-import { TOKEN_STORAGE_KEY } from '@shared/api/config/token-storage-key';
+} from '@shared/lib';
+import { TOKEN_STORAGE_KEY } from '../consts/token-storage-key';
 
 export const setTokenToLocalStorage = setToLocalStorage(TOKEN_STORAGE_KEY);
 export const removeTokenFromLocalStorage =
