--- src/shared/api/model/types.ts
@@ -1,11 +1,0 @@
-export interface ApiResponseTemplate<D = object> {
-  resultCode: 0 | 1;
-  fieldsErrors: string[];
-  messages: string[];
-  data: D;
-}
-
-export interface ApiSuccessResponse<D = object> {
-  data: D;
-  message: string;
-}
