--- src/shared/mui/index.ts
@@ -1,3 +1,0 @@
-export * from './model/types';
-
-export { useMediaQueryState } from './model/useMediaQueryState';
