--- src/shared/redirect/index.ts
@@ -1,1 +1,0 @@
-export { default as useRedirect } from './model/useRedirect';
