--- src/shared/router/index.ts
@@ -1,1 +1,0 @@
-export * as ROUTES from './config/routes';
