--- src/shared/user/model/types.ts
@@ -1,15 +1,0 @@
-export type Photo = string | null;
-
-interface Photos {
-  avatar: Photo;
-  background: Photo;
-}
-
-export interface User {
-  id: number;
-  name: string;
-  job: string;
-  country: string;
-  photos: Photos;
-  followed: boolean;
-}
