--- src/shared/webStorage/index.ts
@@ -1,5 +1,0 @@
-export {
-  getFromLocalStorage,
-  setToLocalStorage,
-  removeFromLocalStorage,
-} from './model/localStorageHandlers';
