+++ src/widgets/navigation/config/profile-items.tsx
@@ -1,7 +1,7 @@
 import PersonIcon from '@mui/icons-material/Person';
 import TurnedInIcon from '@mui/icons-material/TurnedIn';
 import SettingsIcon from '@mui/icons-material/Settings';
-import { ROUTES } from '@shared/router';
+import { ROUTES } from '@shared/consts';
 import { NavigationItem } from '../model/types';
 
 export const PROFILE_NAV_ITEMS: NavigationItem[] = [
