+++ src/widgets/navigation/config/top-items.tsx
@@ -1,6 +1,6 @@
 import NotificationsIcon from '@mui/icons-material/Notifications';
 import EmailIcon from '@mui/icons-material/Email';
-import { ROUTES } from '@shared/router';
+import { ROUTES } from '@shared/consts';
 import { NavigationItem } from '../model/types';
 
 export const TOP_NAV_ITEMS: NavigationItem[] = [
